Created height fade functionality
Started by incorporating leth's heightfade  updates from the following commnet
on github:
https://github.com/brownan/Minecraft-Overviewer/issues/#issue/147/comment/536376

Started experimenting with various fading formulas, and settled on a
logarithmic formula as requested by Irowboat[1] based on Donkey Kong's
adjustments to mcmap[2].

1 - https://github.com/brownan/Minecraft-Overviewer/issues#issue/147/comment/535784
2 - http://www.minecraftforum.net/viewtopic.php?f=25&t=40327&start=30#p772477

Moved the formulas out of chunk.py into heightmap.py and allowed for a way for
a specific function to be added.

Copied Fenixin's formula and code from commit:
https://github.com/Fenixin/Minecraft-Overviewer/commit/28376b8d4f55fc9ed77b9671e43b6507c8e4aff3

I've updated his code to switch automatically switch method of adding white
or black depending on whether a block returned a positive or negative
value for the formula.